refactor: Move enitity related routes from app/main.py to respective modules.
